+++ package.json
@@ -91,10 +91,8 @@
     "react-navigation": "^4.3.9",
     "react-query": "^2.26.2",
     "react-query-devtools": "^2.6.3",
-    "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
     "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
     "request": "^2.88.2",
     "request-promise": "^4.2.5",
     "sequelize": "^4.44.4",
@@ -116,10 +114,8 @@
     "@types/node": "^13.13.9",
     "@types/react": "^16.9.56",
     "@types/react-dom": "^16.9.9",
-    "@types/react-redux": "^7.1.11",
     "@types/react-router-dom": "^5.1.6",
     "@types/react-test-renderer": "^16.9.2",
-    "@types/redux-thunk": "^2.1.0",
     "@types/sequelize": "^4.28.8",
     "@types/validator": "^13.0.0",
     "@types/webpack": "^4.41.25",
