+++ src/App.tsx
@@ -3,8 +3,7 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 import Auth from './routes/auth';
 import Main from './routes/main';
-import { PrivateRoute } from './components/shared/PrivateRoute';
-import { AuthProvider } from './components/shared/AuthProvider';
+import { PrivateRoute, AuthProvider } from './components/Auth';
 
 import './assets/fonts.css';
 import './assets/global.css';
