+++ src/components/Auth/index.ts
@@ -2,3 +2,5 @@ export * from './ResetPassword';
 export * from './ErrorPage';
 export * from './Login';
 export * from './AuthEmail';
+export * from './AuthProvider';
+export * from './PrivateRoute';
