--- src/components/shared/Error404.tsx
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const Error404 = () => <>You at crossroads mate</>;
