+++ src/components/shared/index.ts
@@ -1,6 +1,5 @@
 export * from './Alert';
 export * from './Copyright';
 export * from './LoadingPage';
-export * from './PrivateRoute';
 export * from './TransitionsModal';
 export * from './Error404';
