+++ src/index.tsx
@@ -1,16 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
 
 import App from './app';
-import { store } from './store';
 
-const Index = () => {
-  return (
-    <Provider store={store}>
-      <App />
-    </Provider>
-  );
-};
+const Index = () => <App />;
 
 ReactDOM.render(<Index />, document.getElementById('root'));
