+++ src/routes/auth.tsx
@@ -4,6 +4,7 @@ import { ErrorPage, Login, ResetPassword, AuthEmail } from '../components/Auth';
 
 const Auth = () => {
   const { path } = useRouteMatch();
+  console.log('auth', path);
   return (
     <Switch>
       <Route path={`${path}/login`}>
