+++ src/shared/utilities/index.ts
@@ -1,4 +1,3 @@
 export * from './helperFunctions';
-export * from './useSelector';
 export * from './dates';
 export * from './token';
