--- src/store/index.ts
@@ -1,26 +1,0 @@
-import { createStore, combineReducers, compose, applyMiddleware } from 'redux';
-import ReduxThunk from 'redux-thunk';
-import { authReducer, loadReducer } from './reducers';
-
-declare global {
-  interface Window {
-    __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: typeof compose;
-  }
-}
-
-const rootReducer = combineReducers({
-  auth: authReducer,
-  load: loadReducer,
-});
-
-export type RootState = ReturnType<typeof rootReducer>;
-
-const composeEnhancers =
-  process.env.NODE_ENV !== 'production' && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
-    ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
-    : compose;
-
-export const store = createStore(
-  rootReducer,
-  composeEnhancers(applyMiddleware(ReduxThunk)),
-);
