--- src/store/types/index.ts
@@ -1,3 +1,0 @@
-export * from './authTypes';
-export * from './loadTypes';
-export * from './memberTypes';
